requirements.txt
Django==5.0.3
djangorestframework_simplejwt==5.3.1
matplotlib==3.8.2
